Второй шаг: проверка информации
