+++ js/controllers/level-artist/level-artist-data.js
@@ -1,0 +1,21 @@
+const data = {
+  atist1: {
+    artistId: `1`,
+    artistName: `Твоя Мама`,
+    artistImg: `https://i.ytimg.com/vi/nU4hDsOnCXM/hqdefault.jpg`,
+    artistSong: `./music/Grassy_Hill.mp3`,
+  },
+  atist2: {
+    artistId: `2`,
+    artistName: `Краснознаменная дивизия имени моей бабушки`,
+    artistImg: `https://upload.wikimedia.org/wikipedia/ru/thumb/9/91/80_gvardejskaja_str._divizija.JPG/400px-80_gvardejskaja_str._divizija.JPG`,
+    artistSong: ``,
+  },
+  atist3: {
+    artistId: `3`,
+    artistName: `Те пасоны с гаража`,
+    artistImg: `http://rock-shckola.ru/wp-content/uploads/2015/05/11131.jpg`,
+    artistSong: ``,
+  }
+};
+export default data;
