+++ js/controllers/welcome/welcome-data.js
@@ -1,0 +1,9 @@
+const data = {
+  title: `Правила игры`,
+  text: `Правила просты — за 2 минуты дать
+    максимальное количество правильных ответов.<br>
+    Удачи!`
+};
+
+
+export default data;
